// react/react-primeiro-projeto/src/types/Cliente.ts
export type Cliente = {
    empresa: string;
    filial: string;
    codigo: string;
    loja: string;
    razao: string;
    nome: string;
    tipo: string;
    cnpj: string;
    ie: string;
    pessoa: string;
    email: string;
    cep: string;
    endereco: string;
    bairro: string;
    uf: string;
    municipio: string;
    ddd: string;
    telefone: string;
    tabela: string;
}